fix: no auth config
